delete DBs I'm not supporting.
